# book-network-ui/src/openapi/openapi.json
{
    "openapi": "3.0.1",
    "info": {
          "title": "OpenApi specification - Yusuf",
          "description": "OpenApi documentation for Spring security",
          "termsOfService": "Terms of service",
          "contact": {
                "name": "Yusuf",
                "email": "yusufyuksek513@gmail.com"
          },
          "license": {
                "name": "Licence name",
                "url": "https://some-url.com"
          },
          "version": "1.0"
    },
    "servers": [
          {
                "url": "http://localhost:8088/api/v1",
                "description": "Local ENV"
          },
          {
                "url": "https://some-url.com",
                "description": "Prod ENV"
          }
    ],
    "security": [
          {
                "bearerAuth": [ ]
          }
    ],
    "paths": {
          "/feedbacks": {
                "post": {
                      "tags": [
                            "Feedback"
                      ],
                      "operationId": "saveFeedback",
                      "requestBody": {
                            "content": {
                                  "application/json": {
                                        "schema": {
                                              "$ref": "#/components/schemas/FeedbackRequest"
                                        }
                                  }
                            },
                            "required": true
                      },
                      "responses": {
                            "200": {
                                  "description": "OK",
                                  "content": {
                                        "application/json": {
                                              "schema": {
                                                    "type": "integer",
                                                    "format": "int32"
                                              }
                                        }
                                  }
                            }
                      }
                }
          },
...
          "/feedbacks/book/{book-id}": {
                "get": {
                      "tags": [
                            "Feedback"
                      ],
                      "operationId": "findAllFeedbacksByBook",
                      "parameters": [
                            {
                                  "name": "book-id",
                                  "in": "path",
                                  "required": true,
                                  "schema": {
                                        "type": "integer",
                                        "format": "int32"
                                  }
                            },
                            {
                                  "name": "page",
                                  "in": "query",
                                  "required": false,
                                  "schema": {
                                        "type": "integer",
                                        "format": "int32",
                                        "default": 0
                                  }
                            },
                            {
                                  "name": "size",
                                  "in": "query",
                                  "required": false,
                                  "schema": {
                                        "type": "integer",
                                        "format": "int32",
                                        "default": 10
                                  }
                            }
                      ],
                      "responses": {
                            "200": {
                                  "description": "OK",
                                  "content": {
                                        "application/json": {
                                              "schema": {
                                                    "$ref": "#/components/schemas/PageResponseFeedbackResponse"
                                              }
                                        }
                                  }
                            }
                      }
                }
          },
...
    },
    "components": {
          "schemas": {
                "FeedbackRequest": {
                      "required": [
                            "bookId",
                            "comment"
                      ],
                      "type": "object",
                      "properties": {
                            "note": {
                                  "maximum": 5,
                                  "minimum": 0,
                                  "type": "number",
                                  "format": "double"
                            },
                            "comment": {
                                  "type": "string"
                            },
                            "bookId": {
                                  "type": "integer",
                                  "format": "int32"
                            }
                      }
                },
...
                "FeedbackResponse": {
                      "type": "object",
                      "properties": {
                            "note": {
                                  "type": "number",
                                  "format": "double"
                            },
                            "comment": {
                                  "type": "string"
                            },
                            "ownFeedback": {
                                  "type": "boolean"
                            }
                      }
                },
                "PageResponseFeedbackResponse": {
                      "type": "object",
                      "properties": {
                            "content": {
                                  "type": "array",
                                  "items": {
                                        "$ref": "#/components/schemas/FeedbackResponse"
                                  }
                            },
                            "number": {
                                  "type": "integer",
                                  "format": "int32"
                            },
                            "size": {
                                  "type": "integer",
                                  "format": "int32"
                            },
                            "totalElements": {
                                  "type": "integer",
                                  "format": "int64"
                            },
                            "totalPages": {
                                  "type": "integer",
                                  "format": "int32"
                            },
                            "first": {
                                  "type": "boolean"
                            },
                            "last": {
                                  "type": "boolean"
                            }
                      }
                },
...
          },
          "securitySchemes": {
                "bearerAuth": {
                      "type": "http",
                      "description": "JWT auth description",
                      "in": "header",
                      "scheme": "bearer",
                      "bearerFormat": "JWT"
                }
          }
    }
}